+++ NEWS.md
@@ -1,5 +1,7 @@
 # textrecipes 0.2.1
 
+* `step_word_embeddings()` got a argument `aggregation_default` to specify value in cases where no words matches embedding.
+
 # textrecipes 0.2.0
 
 * `step_tokenize()` got an `engine` argument to specify packages other then tokenizers to tokenize.